+++ Libplanet.Net/Messages/GetChainStatusMsg.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 
 namespace Libplanet.Net.Messages
 {
-    internal class GetChainStatusMsg : Message
+    internal class GetChainStatusMsg : MessageContent
     {
         public override MessageType Type => MessageType.GetChainStatus;